requirements.txt
langchain
streamlit 
langchain_core
langchain_community
langchain_openai
streamlit_chat
python_dotenv
langchain_groq
chainlit
SpeechRecognition
torch
pyaudio
whisper
pygame 
groq
deepgram-sdk==3.0.1
faster_whisper
pydub
audio-recorder-streamlit
faiss-cpu
chromadb
ipykernel
bs4
pypdf
sentence-transformers
langchain-google-genai
langchain-pinecone
langchain-huggingface
pandasql
pandasai
wikipedia
arxiv
langchainhub
langchain_googledrive
googlemaps